Extract location and move comments
